+++ share/templates/js-lib/samples/__name__-sample.jsx
@@ -1,0 +1,11 @@
+import "console.jsx";
+import "{{& filebasename }}.jsx";
+
+class _Main {
+    static function main(argv : string[]) : void
+    {
+        var obj = new {{& Name }}();
+        obj.setMessage("Hello World");
+        console.log(obj.greeting());
+    }
+}