+++ share/templates/js-lib/src/__name__.jsx
@@ -1,0 +1,45 @@
+/**
+ * {{& OrigianlName }}
+ *
+ * {{& description }}
+ *
+ * [sample] This comment is document for class. You can use some HTML tags.
+ *
+ * @author {{& author }}
+ *
+ * @see {{& homepage }}
+ *
+ * License: {{& licensename }}
+ */
+
+__export__ class {{& Name }}
+{
+    var message : string;
+
+    /**
+     * Constructs {{& Name }} object.
+     */
+    function constructor()
+    {
+    }
+
+    /**
+     * [sample] Update message.
+     *
+     * @param message [sample] Message string to store.
+     */
+    function setMessage(message : string) : void
+    {
+        this.message = message;
+    }
+
+    /**
+     * [sample] Returns greeting message.
+     *
+     * @return [sample] greeting message.
+     */
+    function greeting() : string
+    {
+        return this.message;
+    }
+}