+++ share/templates/js-lib/tmp/lib/tmp.js
@@ -1,0 +1,13 @@
+/*
+ * tmp
+ * https://github.com/yshibukawa/tmp
+ *
+ * Copyright (c) 2013 Yoshiki Shibukawa
+ * Licensed under the MIT license.
+ */
+
+'use strict';
+
+exports.awesome = function() {
+  return 'awesome';
+};